requirements.txt
streamlit==1.31.0
openai==1.51.0
httpx==0.27.0
pydantic==2.6.1
PyMuPDF==1.23.21
python-docx==1.1.0
pytesseract==0.3.10
sentence-transformers==2.3.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
email-validator==2.1.0
phonenumbers==8.13.28
pillow==10.2.0
plotly==5.18.0
altair==5.2.0
jsonschema==4.21.1
rapidfuzz==3.6.1
langdetect==1.0.9
python-dotenv==1.0.0
